\name{GenABEL.data-package}
\alias{GenABEL.data-package}
\alias{GenABEL.data}
\docType{package}
\title{
GenABEL.data: An R package for storing GenABEL data
}
\description{
GenABEL.data contains six files with data which is used by examples of GenABEL. 
These are ge03d2.clean.RData, ge03d2c.RData, ge03d2ex.clean.RData, ge03d2ex.RData, ge03d2.RData and srdta.RData.
The files are the same as they were GenABEL package. 

}
\details{
\tabular{ll}{
Package: \tab GenABEL.data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-27\cr
License: \tab GPL (>= 2)\cr
}
GenABEL.data package can be used only together with GenABEL package. 
}
\author{
Authors: Maksim Struchalin <m.v.struchalin@mail.ru> and GenABEL project developers <genabel.project@gmail.com>
}
\references{
www.genabel.org
}
\keyword{ package }
\seealso{
GenABEL
}
\examples{
data(srdta)
}
